backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "google-auth>=2.49.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pyseto>=1.9.1",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "supabase>=2.28.2",
    "uvicorn>=0.42.0",
]
